chore() rename swiftui view and add events
